Module DeeplangType

module ParseTree : sig ... end

This module defines the abstract syntax of deeplang. The syntax tree will be used for all subsequent compiler passes, until code generation.

module Walker : sig ... end

This module walks AST from ParseTree, to do name resolution and type checking in one pass.