Module DeeplangType.Walker

This module walks AST from ParseTree, to do name resolution and type checking in one pass.

Programmer should read from the Interface.

Types

module NameMap : sig ... end

type name_map = ParseTree.symbol NameMap.t

type context = {

mutable name_map : name_map;

}

type layered_name_map = {

mutable deep : name_map;

mutable shallow : name_map;	(*
helps to forbide variable with the same names in one scope
*)

}

TODO:layered_name_map

type var_data = {

mut : ParseTree.mutability;

typ : ParseTree.typ;

name : ParseTree.variable;

}

type var_table = ( ParseTree.symbol, var_data ) Stdlib.Hashtbl.t

type fun_data = {

args : ParseTree.typ list;

retv : ParseTree.typ;

name : ParseTree.variable list;

}

type fun_table = ( ParseTree.func_name, fun_data ) Stdlib.Hashtbl.t

type adty_data = {

mutable intf : ParseTree.intf_name list;

mutable meth : fun_table;

core : (ParseTree.adt_label * ParseTree.typ list) list;

}

type is_delegated =

| IsDelegated

| JustMember

type field_data = {

typ : ParseTree.typ;

delegated : is_delegated;

}

type struct_data = {

mutable intf : ParseTree.intf_name list;

mutable meth : fun_table;

core : ( ParseTree.struct_field, field_data ) Stdlib.Hashtbl.t;

fields : ParseTree.struct_def_field list;

}

type intf_data = {

meth : fun_table;

}

type typ_data =

| ADT_data of adty_data

| Struct_data of struct_data

| Intf_data of intf_data

type typ_table = ( ParseTree.typ_name, typ_data ) Stdlib.Hashtbl.t

type adt_data = {

sum : ParseTree.typ_name;	(*
sum_type's name
*)

typ : ParseTree.typ list;	(*
parameters of constructor
*)

}

type adt_table = ( ParseTree.adt_label, adt_data ) Stdlib.Hashtbl.t

type ref_data = {

sym : ParseTree.symbol;	(*
point to the varaible
*)

}

type ref_table = ( ParseTree.NodeId.expr, ref_data ) Stdlib.Hashtbl.t

type table = {

var : var_table;	(*
the mutability, type and name of variable
*)

fnc : fun_table;	(*
the signature/type of function
*)

typ : typ_table;	(*
the info of adt, struct and interface
*)

adt : adt_table;	(*
the signature/type of adt_labels/constructor
*)

ref : ref_table;	(*
the use of variable
*)

}

module BuiltinType : sig ... end

BuiltinType used in walker, which is not yet defined in ParsingTree

Implementation

val walk_add_variable : ParseTree.symbol -> ParseTree.mutability -> ParseTree.variable -> ParseTree.typ option -> table -> layered_name_map -> unit

add variable in table and name_map.deep, .shallow

val walk_pattern : ParseTree.pattern -> ParseTree.typ -> table -> layered_name_map -> unit

call walk_add_variable

# walk_pattern pattern typ table layered_name_map

typ is the type of expr corresponding to pattern

In let (x, _) = (3, 1.1) , pattern is (x, _), and typ is (i64, f64).

val walk_expr : ParseTree.expr -> table -> name_map -> ParseTree.typ -> ParseTree.typ

Type Check implemented in walk_expr:

# let ty = walk_expr expr table name_map type_this 
  val ty : typ = <...>

type_this is the type of this when in method scope.

ty is the type of expr.

val walk_stmt : ParseTree.stmt -> table -> layered_name_map -> ParseTree.typ -> ParseTree.typ option -> ParseTree.typ option

walk stmt with checking the return type

# let typo' = walk_stmt stmt table layered_name_map type_this typo
  val typo' : typ option = <...>

typo and typo' are the return type of expr in return expr.

match typo with
| None     -> have walked no          return-stmts yet
| Some(ty) -> have walked one or more return-stmts and
              their return types are all ty

When meeting a return-stmt, check typo to ensure unique return type.

val walk_func_decl : ParseTree.func_name -> ParseTree.func_arg list -> ParseTree.typ -> table -> unit

check func name, and collect type info

# walk_func_decl name parameters return_type table

val walk_func_stmt : ParseTree.typ -> ParseTree.func_arg list -> ParseTree.stmt -> table -> name_map -> ParseTree.typ option

simply walk function's body inductively

val walk_method_intf : ParseTree.intf_name option -> ParseTree.typ_name -> table -> unit

impl intf for T with methods, check whether intf exists and whether it has such methods

val walk_method_decl : ParseTree.typ_name -> ParseTree.func_name -> ParseTree.func_arg list -> ParseTree.typ -> table -> unit

similar to walk_func_decl, but checked in the scope of type T

val walk_method_stmt : ParseTree.typ -> ParseTree.func_arg list -> ParseTree.stmt -> table -> name_map -> ParseTree.typ option

val walk_top : ParseTree.top_clause -> table -> name_map -> name_map

Interface

val walk : context -> table -> ParseTree.top_clause -> unit

context is mutable, mapping variable names to symbols.

table is mutable, produced by walk.

Indeed, walk is just a wrap of walk_top, see the code.

TODO:Test Helper

val mk_id : unit -> int

val test_span : ParseTree.src_span

val top_clause_gen : ParseTree.top_clause_shape -> ParseTree.top_clause

val expr_gen : ParseTree.expr_shape -> ParseTree.expr

val test_case : ParseTree.top_clause

val context : context

val table : table