Module DeeplangType__ParseTree.NodeId

NodeId contains types for unique ids for various kinds of AST nodes. Can be used to attach extra data to AST nodes, via an external table. DISCUSSION WANTED:

separate variable usage ids and node ids?

type typ_var =

| TVarId of int

type typ_name =

| TypNameId of int

NodeId.typ_var and NodeId.typ_name are used for scope resolution of type variables and type names.

type pattern =

| PatId of int

type expr =

| ExprId of int

type stmt =

| StmtId of int

type intf_impl =

| IntfImplId of int

type func_impl =

| FuncImplId of int